Set electron name to 'Parity-UI'
Electron stores the user data by default in `~/.config/<app name>`

The app name is retrieved from the package.json when Electron is
run on a folder (e.g. `electron .`); default app name 'Electron'
is used if Electron is run directly on a script, as is the case.

Electron can be run either on .build/ or on .electron/ (dev),
so it's not practical to use package.json to set the app name
(.electron/ has no package.json); hence it's better in our case
to define it in the code.